added Contracts models and contracts controller

+++ SystemHRApi/Models/Employees.cs
@@ -2,12 +2,12 @@
 {
     public class Employees
     {
-        public int Id { get; set; }
-        public string? Name { get; set; }    
+        public long Id { get; set; }
+        public string? Name { get; set; }
         public string? Surname { get; set; }
         public string? Adress { get; set; }
-        public int NumberPhone { get; set; }
-        public  string? DateOfBirth { get; set; }
+        public long NumberPhone { get; set; }
+        public string? DateOfBirth { get; set; }
         public string? DateOfEmployment { get; set; }
         public string? Contract { get; set; }
     }
