+++ SystemHRApi/ModelsDTO/EmployeesDTO.cs
@@ -2,11 +2,11 @@
 {
     public class EmployeesDTO
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string? Name { get; set; }
         public string? Surname { get; set; }
         public string? Adress { get; set; }
-        public int NumberPhone { get; set; }
+        public long NumberPhone { get; set; }
         public string? DateOfBirth { get; set; }
         public string? DateOfEmployment { get; set; }
         public string? Contract { get; set; }
